+++ cookbooks/arcgis-enterprise/attributes/server.rb
@@ -33,11 +33,19 @@ default['arcgis']['server'].tap do |server|
   server['private_url'] = 'https://' + node['arcgis']['server']['domain_name'] + ':6443/arcgis'
   server['web_context_url'] = 'https://' + node['arcgis']['server']['domain_name'] + '/' + node['arcgis']['server']['wa_name']
   server['admin_username'] = 'admin'
-  server['admin_password'] = 'changeit'
+  if ENV['ARCGIS_SERVER_ADMIN_PASSWORD'].nil?
+    server['admin_password'] = 'changeit'
+  else
+    server['admin_password'] = ENV['ARCGIS_SERVER_ADMIN_PASSWORD']
+  end
   server['managed_database'] = ''
   server['replicated_database'] = ''
   server['keystore_file'] = ''
-  server['keystore_password'] = ''
+  if ENV['ARCGIS_SERVER_KEYSTORE_PASSWORD'].nil?
+    server['keystore_password'] = ''
+  else
+    server['keystore_password'] = ENV['ARCGIS_SERVER_KEYSTORE_PASSWORD']
+  end
   server['cert_alias'] = node['arcgis']['server']['domain_name']
   server['system_properties'] = {}
   server['log_level'] = 'WARNING'
@@ -46,6 +54,7 @@ default['arcgis']['server'].tap do |server|
   server['configure_autostart'] = true
   server['install_system_requirements'] = true
   server['use_join_site_tool'] = false
+  server['soc_max_heap_size'] = 64
 
   unless node['arcgis']['server']['authorization_file'].nil?
     server['cached_authorization_file'] = ::File.join(Chef::Config[:file_cache_path],
@@ -60,6 +69,10 @@ default['arcgis']['server'].tap do |server|
 
   server['authorization_retries'] = 10
 
+  server['security']['user_store_config'] = {'type' => 'BUILTIN', 'properties' => {}}
+  server['security']['role_store_config'] = {'type' => 'BUILTIN', 'properties' => {}}
+  server['security']['privileges'] = {'PUBLISH' => [], 'ADMINISTER' => []}
+
   case node['platform']
   when 'windows'
     server['authorization_tool'] = ENV['ProgramW6432'] + '\\Common Files\\ArcGIS\\bin\\SoftwareAuthorization.exe'
@@ -70,17 +83,37 @@ default['arcgis']['server'].tap do |server|
                                   'ArcGISServer', 'Setup.exe')
     server['lp-setup'] = node['arcgis']['server']['setup']
     server['install_dir'] = ENV['ProgramW6432'] + '\\ArcGIS\\Server'
+
     server['local_directories_root'] = 'C:\\arcgisserver'
-    server['directories_root'] = node['arcgis']['server']['local_directories_root']
-    server['log_dir'] = ::File.join(node['arcgis']['server']['local_directories_root'], 'logs')
+
+    if node['arcgis']['server']['local_directories_root'].nil?
+      server['directories_root'] = server['local_directories_root']
+    else
+      server['directories_root'] = node['arcgis']['server']['local_directories_root']
+    end
+
+    if node['arcgis']['server']['local_directories_root'].nil?
+      server['log_dir'] = ::File.join(server['local_directories_root'], 'logs')
+    else
+      server['log_dir'] = ::File.join(node['arcgis']['server']['local_directories_root'], 'logs') 
+    end
+
     server['config_store_type'] = 'FILESYSTEM'
-    server['config_store_connection_string'] = ::File.join(node['arcgis']['server']['directories_root'], 'config-store')
+    if node['arcgis']['server']['directories_root'].nil?
+      server['config_store_connection_string'] = ::File.join(server['directories_root'], 'config-store')
+    else
+      server['config_store_connection_string'] = ::File.join(node['arcgis']['server']['directories_root'], 'config-store')
+    end
     server['config_store_connection_secret'] = ''
 
     case node['arcgis']['version']
+    when '10.6'
+      server['setup_archive'] = ::File.join(node['arcgis']['repository']['archives'],
+                                            'ArcGIS_Server_Windows_106_159940.exe')
+      server['product_code'] = '{07606F78-D997-43AE-A9DC-0738D91E8D02}'
     when '10.5.1'
       server['setup_archive'] = ::File.join(node['arcgis']['repository']['archives'],
-                                            'ArcGIS_Server_Windows_XXXXXX.exe')
+                                            'ArcGIS_Server_Windows_1051_156124.exe')
       server['product_code'] = '{40CC6E89-93A4-4D87-A3FB-11413C218D2C}'
     when '10.5'
       server['setup_archive'] = ::File.join(node['arcgis']['repository']['archives'],
@@ -116,9 +149,12 @@ default['arcgis']['server'].tap do |server|
     server['lp-setup'] = node['arcgis']['server']['setup']
 
     case node['arcgis']['version']
+    when '10.6'
+        server['setup_archive'] = ::File.join(node['arcgis']['repository']['archives'],
+                                              'ArcGIS_Server_Linux_106_159989.tar.gz')
     when '10.5.1'
         server['setup_archive'] = ::File.join(node['arcgis']['repository']['archives'],
-                                              'ArcGIS_Server_Linux_1051_XXXXXX.tar.gz')
+                                              'ArcGIS_Server_Linux_1051_156429.tar.gz')
     when '10.5'
       server['setup_archive'] = ::File.join(node['arcgis']['repository']['archives'],
                                             'ArcGIS_Server_Linux_105_154052.tar.gz')