+++ cookbooks/arcgis-enterprise/resources/server.rb
@@ -19,7 +19,8 @@
 
 actions :system, :unpack, :install, :uninstall, :update_account, :stop, :start,
         :configure_autostart, :authorize, :create_site, :join_site,
-        :join_cluster, :configure_https, :register_database, :federate
+        :join_cluster, :configure_https, :register_database, :federate,
+        :set_identity_store, :assign_privileges, :set_machine_properties
 
 attribute :setup_archive, :kind_of => String
 attribute :setups_repo, :kind_of => String
@@ -58,6 +59,11 @@ attribute :log_level, :kind_of => String, :default => 'WARNING'
 attribute :log_dir, :kind_of => String
 attribute :max_log_file_age, :kind_of => Integer, :default => 90
 attribute :use_join_site_tool, :kind_of => [TrueClass, FalseClass], :default => false
+attribute :user_store_config, :kind_of => Hash, :default => {}
+attribute :role_store_config, :kind_of => Hash, :default => {}
+attribute :privileges, :kind_of => Hash, :default => {}
+attribute :soc_max_heap_size, :kind_of => Integer, :default => 64
+
 def initialize(*args)
   super
   @action = :install